batch/requirements.txt
policyengine-core>=3.0.0,<3.23.0
pandas>=2.0.0
numpy>=1.24.0
google-cloud-storage>=2.10.0
tqdm>=4.65.0
